Schema
Remove passenger_id from Bookings and instead add Booking_reference to Bookings and add Booking_id to passengers. The reason for this was that bookings did not need a reference to passengers. Passengers instead should have a booking_id, and passengers can use a booking reference to search for bookings. This change matches the new belongs to relationships i.e. passengers belong to bookings, not the other way round as it was originally created