Stimulus
Add remove controller with Connect action. The data-controller attribute is set to the div encompassing the passenger form, so applies to all instances of the passenger that gets created. The remove button for each form is set to the target, and using buttonTarget in Stimulus only targets the first occurrence of this target, so makes the button disappear and disabled for the first passenger element